refactor: 캐싱 방식 수정
